requirements.txt
concrete-python==2.10.0
filelock==3.18.0
fsspec==2025.3.2
importlib_resources==6.5.2
Jinja2==3.1.6
jsonpickle==4.0.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
scipy==1.15.2
sympy==1.13.3
torch==2.2.2
typing_extensions==4.13.2
z3-solver==4.13.0.0
